requirements.txt
python-dotenv>=1.0.0
psycopg[binary]>=3.0.0
spacy>=3.0.0
boto3>=1.26.0
markdown>=3.4.0
beautifulsoup4>=4.12.0 